Fixed bug that created additional lexer action entries.
The subtleties wrt. to object equality vs. reference equality between Java and TypeScript are a constant source of tricky bugs. However, there's not need to always use a OrderedHashMap if the key is a standard type (a type without equals/hashMap methods) as the standard Map already maintains insertion order.